fix: update slippageBps to double type in SwapViewModel

+++ AndroidDemo/app/src/main/java/io/ton/walletkit/demo/presentation/viewmodel/SwapViewModel.kt
@@ -160,8 +160,8 @@ class SwapViewModel(
                         from = current.fromToken,
                         to = current.toToken,
                         network = TONNetwork.MAINNET,
-                        slippageBps = current.slippageBps,
-                        maxOutgoingMessages = 4,
+                        slippageBps = current.slippageBps.toDouble(),
+                        maxOutgoingMessages = 4.0,
                         isReverseSwap = current.isReverseSwap,
                     ),
                 )
